refactor: Curation cleanup
